requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==21.2.0
Automat==0.6.0
beautifulsoup4==4.9.3
blinker==1.4
certifi==2018.1.18
chardet==4.0.0
click==6.7
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
discord.py==1.7.2
distro-info===0.18ubuntu0.18.04.1
Django==3.2.3
django-bootstrap4==3.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==3.2
idna-ssl==1.1.0
importlib-metadata==2.1.1
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
multidict==5.1.0
netifaces==0.10.4
oauthlib==2.0.6
PAM==0.4.2
psutil==5.8.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-debian==0.1.32
python-dotenv==0.17.1
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
soupsieve==2.2.1
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
tinydb==4.4.0
Twisted==17.9.0
typing-extensions==3.10.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
yarl==1.6.3
zipp==3.4.1
zope.interface==4.3.2
